refactor: 파일 위치, 라우터 재정의
