+++ client/src/Router.jsx
@@ -1,21 +1,18 @@
 import { Routes, Route } from "react-router-dom";
 import LoginPage from "./pages/LoginPage";
 import MainPage from "./pages/MainPage";
-import ProfilePage from "./pages/ProfilePage";
 import PrivateRoute from "./components/PrivateRoute";
 
-import AccountDetailPage from './pages/NavBarControls/AccountDetail';
-import Setting from './pages/NavBarControls/Setting';
-import HelpCenter from './pages/NavBarControls/HelpCenter';
-import ContactUs from "./pages/NavBarControls/ContactUs";
+import ProfilePage from './pages//Profile/ProfilePage';
+import HelpCenter from './pages/Profile/HelpCenter';
+import ContactUs from "./pages/Profile/ContactUs";
 
 export default function Router() {
     return (
         <Routes>
             <Route path="/login" element={<LoginPage />} />
 
-            <Route path="/AccountDetail" element={<AccountDetailPage />} />
-            <Route path="/Setting" element={<Setting />} />
+            <Route path="/ProfilePage" element={<ProfilePage />} />
             <Route path="/HelpCenter" element={<HelpCenter />} />
             <Route path="/ContactUs" element={<ContactUs />} />
